Diffusion Pseudo Time

Create pseudotime ordering and assigns cell to one of three branches

DPT(dm, tips = random_root(dm), ..., w_width = 0.1)

Arguments

dm
A DiffusionMap object. Its transition probabilities will be used to calculate the DPT

tips
The cell index/indices from which to calculate the DPT(s) (integer of length 1-3)

...
Unused. All parameters to the right of the ... have to be specified by name (e.g. DPT(dm, w_width = 0.2))

w_width
Window width to use for deciding the branch cutoff

Value

A DPT object:

Details

Treat it as a matrix of pseudotime by subsetting ([ dim nrow ncol as.matrix), and as a list of pseudodime, and expression vectors ($ [[ names as.data.frame).

Slots

branch

matrix (of integer) recursive branch labels for each cell (row); NA for undeceided. Use branch_divide to modify this.

tips

matrix (of logical) indicating if a cell (row) is a tip of the corresponding banch level (col)

dm

DiffusionMap used to create this DPT object

Examples

data(guo_norm)
dm <- DiffusionMap(guo_norm)
dpt <- DPT(dm)
str(dpt)

#> Formal class 'DPT' [package "destiny"] with 3 slots
#>   ..@ branch: int [1:428, 1:7] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:7] "Branch1" "Branch2" "Branch3" "Branch4" ...
#>   ..@ tips  : logi [1:428, 1:7] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. ..- attr(*, "dimnames")=List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:7] "Tips1" "Tips2" "Tips3" "Tips4" ...
#>   ..@ dm    :Formal class 'DiffusionMap' [package "destiny"] with 17 slots
#>   .. .. ..@ eigenvalues  : num [1:20] 0.947 0.807 0.724 0.712 0.654 ...
#>   .. .. ..@ eigenvectors : num [1:428, 1:20] 0.0483 0.0468 0.0401 0.045 0.0479 ...
#>   .. .. .. ..- attr(*, "dimnames")=List of 2
#>   .. .. .. .. ..$ : NULL
#>   .. .. .. .. ..$ : chr [1:20] "DC1" "DC2" "DC3" "DC4" ...
#>   .. .. ..@ sigmas       :Formal class 'Sigmas' [package "destiny"] with 5 slots
#>   .. .. .. .. ..@ log_sigmas   : NULL
#>   .. .. .. .. ..@ dim_norms    : NULL
#>   .. .. .. .. ..@ optimal_sigma: num [1:428] 9.72 9.28 10.9 9.69 9.52 ...
#>   .. .. .. .. ..@ optimal_idx  : NULL
#>   .. .. .. .. ..@ avrd_norms   : NULL
#>   .. .. ..@ data_env     :<environment: 0x55943d27aa48> 
#>   .. .. ..@ eigenvec0    : num [1:428] 0.0546 0.0535 0.0513 0.0529 0.053 ...
#>   .. .. ..@ transitions  :Formal class 'dsCMatrix' [package "Matrix"] with 7 slots
#>   .. .. .. .. ..@ i       : int [1:91378] 0 0 1 0 1 2 0 1 2 3 ...
#>   .. .. .. .. ..@ p       : int [1:429] 0 0 1 3 6 10 15 21 28 36 ...
#>   .. .. .. .. ..@ Dim     : int [1:2] 428 428
#>   .. .. .. .. ..@ Dimnames:List of 2
#>   .. .. .. .. .. ..$ : NULL
#>   .. .. .. .. .. ..$ : NULL
#>   .. .. .. .. ..@ x       : num [1:91378] 0.1043 0.0445 0.0301 0.0413 0.039 ...
#>   .. .. .. .. ..@ uplo    : chr "U"
#>   .. .. .. .. ..@ factors : list()
#>   .. .. ..@ d            : num [1:428] 3.76 4.05 4.86 4.82 3.5 ...
#>   .. .. ..@ d_norm       : num [1:428] 0.167 0.161 0.148 0.157 0.158 ...
#>   .. .. ..@ k            : int 427
#>   .. .. ..@ n_local      : int [1:3] 5 6 7
#>   .. .. ..@ density_norm : logi TRUE
#>   .. .. ..@ rotate       : logi FALSE
#>   .. .. ..@ distance     : chr "euclidean"
#>   .. .. ..@ censor_val   : NULL
#>   .. .. ..@ censor_range : NULL
#>   .. .. ..@ missing_range: NULL
#>   .. .. ..@ vars         : NULL