Sigmas Object

Holds the information about how the sigma parameter for a DiffusionMap was obtained, and in this way provides a plotting function for the find_sigmas heuristic. You should not need to create a Sigmas object yourself. Provide sigma to DiffusionMap instead or use find_sigmas.

Sigmas(...)

# S4 method for Sigmas
optimal_sigma(object)

# S4 method for Sigmas
print(x)

# S4 method for Sigmas
show(object)

Arguments

object, x
Sigmas object

...
See “Slots” below

Value

Sigmas creates an object of the same class

optimal_sigma retrieves the numeric value of the optimal sigma or local sigmas

Details

A Sigmas object is either created by find_sigmas or by specifying the sigma parameter to DiffusionMap.

In the second case, if the sigma parameter is just a number, the resulting Sigmas object has all slots except of optimal_sigma set to NULL.

Slots

log_sigmas

Vector of length \(m\) containing the \(\log_{10}\) of the \(\sigma\)s

dim_norms

Vector of length \(m-1\) containing the average dimensionality \(\langle p \rangle\) for the respective kernel widths

optimal_sigma

Multiple local sigmas or the mean of the two global \(\sigma\)s around the highest \(\langle p \rangle\) (c(optimal_idx, optimal_idx+1L))

optimal_idx

The index of the highest \(\langle p \rangle\).

avrd_norms

Vector of length \(m\) containing the average dimensionality for the corresponding sigma.

See also

Examples

data(guo)
sigs <- find_sigmas(guo, verbose = FALSE)
optimal_sigma(sigs)

#> [1] 7.663719

print(sigs)

#> Sigmas (10 Steps performed)
#> optimal_sigma:  num 7.66