Compatibility versions of functions for destiny 1.x scripts

This is most useful if you have old scripts that use the old argument order and the default being global sigma. Simply put DiffusionMap <- DiffusionMapCompat on top of your script to make it work with destiny 2.x.

DiffusionMapCompat(data, sigma = "global", k = find_dm_k(nrow(data) -
  1L), n.eigs = min(20L, nrow(data) - 2L), density.norm = TRUE, ...,
  distance = c("euclidean", "cosine", "rankcor"), censor.val = NULL,
  censor.range = NULL, missing.range = NULL, vars = NULL,
  verbose = !is.null(censor.range))

cube.helix(n = 6, start = 0, r = 0.4, hue = 0.8, gamma = 1,
  light = 0.85, dark = 0.15, reverse = FALSE)

dm.predict(dm, new.data, verbose = FALSE)

eig.decomp(M, n.eigs, sym = isSymmetric(M))

find.dm.k(n, min.k = 100L, small = 1000L, big = 10000L)

find.sigmas(data, step.size = 0.1, steps = 10L, start = NULL,
  sample.rows = 500L, early.exit = FALSE, ..., censor.val = NULL,
  censor.range = NULL, missing.range = NULL, vars = NULL,
  verbose = TRUE)

optimal.sigma(object)

lWhich(idx, nms = seq_len(len), len = length(nms), useNames = TRUE)

plot_gradient_map(coords, exprs, ..., gene, dims = 1:2,
  pal = hcl.colors, faceter = facet_wrap(~Gene))

Arguments

data, n, n.eigs, start, censor.val, censor.range, missing.range, vars, verbose, ...
Corresponding parameters in multiple functions

sigma, k, density.norm, distance
Corresponding DiffusionMap parameters

r, hue, gamma, light, dark, reverse
Corresponding cube_helix parameters

dm, new.data
Corresponding dm_predict parameters

M, sym
Corresponding eig_decomp parameters

min.k, small, big
Corresponding find_dm_k parameters

step.size, steps, sample.rows, early.exit
Corresponding find_sigmas parameters

object
Corresponding optimal_sigma parameters

Examples

DiffusionMap <- DiffusionMapCompat
# NOT RUN {
source('my-old-script.r')
# }